Unreviewed, reverting 301450@main (0471d4f78def) https://bugs.webkit.org/show_bug.cgi?id=300722 rdar://162627041
Broke API test TestWTF.WTF_OptionSet.AddAndRemove

Reverted change:

    Replace PseudoIdSet with EnumSet<PseudoId>
    https://bugs.webkit.org/show_bug.cgi?id=300396
    rdar://162214153
    301450@main (0471d4f78def)

Canonical link: https://commits.webkit.org/301494@main

+++ Source/WebCore/animation/WebAnimationUtilities.h
@@ -33,7 +33,7 @@
 
 namespace WebCore {
 
-enum class PseudoId : uint8_t;
+enum class PseudoId : uint32_t;
 
 class AnimationEventBase;
 class Document;
